+++ hs/app/Main.hs
@@ -13,6 +13,10 @@ import qualified HelVM.HelMA.Automata.BrainFuck.Lexer      as BF
 import qualified HelVM.HelMA.Automata.ETA.Automaton        as ETA
 import qualified HelVM.HelMA.Automata.ETA.Lexer            as ETA
 
+import qualified HelVM.HelMA.Automata.LazyK.API.ParserType as LK
+import qualified HelVM.HelMA.Automata.LazyK.Automaton      as LK
+
+
 import qualified HelVM.HelMA.Automata.SubLeq.Automaton     as SQ
 import qualified HelVM.HelMA.Automata.SubLeq.Lexer         as SQ
 
@@ -22,8 +26,9 @@ import qualified HelVM.HelMA.Automata.WhiteSpace.Parser    as WS
 
 import qualified HelVM.HelMA.Automata.Zot.Automaton        as Zot
 
-import           HelVM.HelMA.Automaton.API.EvalParams
+
 import           HelVM.HelMA.Automaton.API.IOTypes
+import           HelVM.HelMA.Automaton.API.RunParams
 import           HelVM.HelMA.Automaton.API.TypeOptions
 
 import           HelVM.HelMA.Automaton.IO.BusinessIO
@@ -51,21 +56,21 @@ main = runApp =<< execParser opts where
      <> progDesc "Runs esoteric programs - complete with pretty bad error messages" )
 
 runApp:: AppOptions -> IO ()
-runApp (AppOptions lang visibleTokens minified emitTL emitIL printLogs compile asciiLabels ramType stackType cellType intCellType dumpType exec file) = do
+runApp (AppOptions lang visibleTokens parserType minified emitTL emitIL printLogs compile asciiLabels ramType stackType cellType intCellType dumpType exec file) = do
   hSetBuffering stdout IO.NoBuffering
   source <- readSourceFile exec file
-  run minified emitTL emitIL printLogs typeOptions asciiLabels compile (parseLang lang) (enumFromBool visibleTokens) source
+  run minified emitTL emitIL printLogs typeOptions asciiLabels compile (parseLang lang) (enumFromBool visibleTokens) (LK.parseParserType parserType) source
     where typeOptions = TypeOptions (parseRAMType ramType) (parseStackType stackType) (parseCellType cellType) (parseIntCellType intCellType) (parseDumpType dumpType)
 
 readSourceFile :: Exec -> String -> IO Source
 readSourceFile True = pure . toText
 readSourceFile _    = readFileText
 
-run :: Minified -> EmitTL -> EmitIL -> PrintLogs -> TypeOptions -> Compile -> AsciiLabels -> Lang -> TokenType -> Source -> IO ()
-run True _    _    _ _ _ _ l t s  = minification l t s
-run _    True _    _ _ _ _ l t s  = tokenize l t s
-run _    _    True _ _ _ a l t s  = parse l t a s
-run _    _    _    p to c a l t s = eval p to c a l t s
+run :: Minified -> EmitTL -> EmitIL -> PrintLogs -> TypeOptions -> Compile -> AsciiLabels -> Lang -> TokenType -> LK.ParserType -> Source -> IO ()
+run True _    _    _ _ _ _ l t _ s = minification l t s
+run _    True _    _ _ _ _ l t _ s = tokenize l t s
+run _    _    True _ _ _ a l t _ s = parse l t a s
+run _    _    _    p o c a l t k s = eval p o c a l t k s
 
 minification :: Lang -> TokenType -> Source -> IO ()
 minification WS VisibleTokenType = print . WS.readVisibleTokens
@@ -87,15 +92,16 @@ parse WS   VisibleTokenType a = pPrintNoColor . WS.flipParseVisible a
 parse WS   WhiteTokenType   a = pPrintNoColor . WS.flipParseWhite   a
 parse lang tt               _ = tokenize lang tt
 
-eval :: PrintLogs -> TypeOptions -> Compile -> AsciiLabels -> Lang -> TokenType -> Source -> IO ()
-eval p options c a lang tt s = (controlTToIO p . evalParams lang tt) params
-  where params = EvalParams {compile = c , asciiLabel = a , source = s , typeOptions = options}
+eval :: PrintLogs -> TypeOptions -> Compile -> AsciiLabels -> Lang -> TokenType -> LK.ParserType -> Source -> IO ()
+eval p options c a lang tt pt s = (controlTToIO p . runWithParams lang tt pt) params
+  where params = RunParams {compile = c , asciiLabel = a , source = s , typeOptions = options}
 
-evalParams :: BIO m => Lang -> TokenType -> EvalParams -> m ()
-evalParams WS tt = WS.evalParams tt
-evalParams Cat _ = Cat.evalParams
-evalParams Rev _ = Rev.evalParams
-evalParams BF  _ = BF.evalParams
-evalParams ETA _ = ETA.evalParams
-evalParams SQ  _ = SQ.evalParams
-evalParams Zot _ = Zot.evalParams
+runWithParams :: BIO m => Lang -> TokenType -> LK.ParserType -> RunParams -> m ()
+runWithParams LK  _ p = LK.runWithParams p
+runWithParams WS  t _ = WS.runWithParams t
+runWithParams Cat _ _ = Cat.runWithParams
+runWithParams Rev _ _ = Rev.runWithParams
+runWithParams BF  _ _ = BF.runWithParams
+runWithParams ETA _ _ = ETA.runWithParams
+runWithParams SQ  _ _ = SQ.runWithParams
+runWithParams Zot _ _ = Zot.runWithParams