+++ hs/src/HelVM/HelMA/Automata/WhiteSpace/Automaton.hs
@@ -1,9 +1,9 @@
 module HelVM.HelMA.Automata.WhiteSpace.Automaton (
-  simpleEval,
-  evalParams,
-  eval,
-  evalIL,
-  evalTL,
+  simpleRun,
+  runWithParams,
+  run,
+  runIL,
+  runTL,
   start,
 ) where
 
@@ -13,8 +13,8 @@ import           HelVM.HelMA.Automata.WhiteSpace.Parser
 import           HelVM.HelMA.Automata.WhiteSpace.Symbol
 import           HelVM.HelMA.Automata.WhiteSpace.Token
 
-import           HelVM.HelMA.Automaton.API.EvalParams
 import           HelVM.HelMA.Automaton.API.IOTypes
+import           HelVM.HelMA.Automaton.API.RunParams
 
 import           HelVM.HelMA.Automaton.IO.BusinessIO
 import           HelVM.HelMA.Automaton.IO.EvaluatorIO
@@ -40,30 +40,30 @@ import qualified HelVM.HelIO.Collections.SList                as SList
 import qualified Data.Sequence                                as Seq
 import qualified Data.Vector                                  as Vector
 
-simpleEval :: BIO m => S.SimpleParams -> m ()
-simpleEval p = eval (S.tokenType p) (S.source p) (S.asciiLabel p) (S.stackType p) (S.ramType p) (S.dumpType p)
+simpleRun :: BIO m => S.SimpleParams -> m ()
+simpleRun p = run (S.tokenType p) (S.source p) (S.asciiLabel p) (S.stackType p) (S.ramType p) (S.dumpType p)
 
 ----
 
-evalParams :: BIO m => TokenType -> EvalParams -> m ()
-evalParams tokenType p = eval tokenType (source p) (asciiLabel p) (stackTypeOptions p) (ramTypeOptions p) (dumpTypeOptions p)
+runWithParams :: BIO m => TokenType -> RunParams -> m ()
+runWithParams tokenType p = run tokenType (source p) (asciiLabel p) (stackTypeOptions p) (ramTypeOptions p) (dumpTypeOptions p)
 
-eval :: BIO m => TokenType -> Source -> Bool -> StackType -> RAMType -> DumpType -> m ()
-eval tokenType source = evalTL $ tokenize tokenType source
+run :: BIO m => TokenType -> Source -> Bool -> StackType -> RAMType -> DumpType -> m ()
+run tokenType source = runTL $ tokenize tokenType source
 
-evalTL :: BIO m => TokenList -> Bool -> StackType -> RAMType -> DumpType -> m ()
-evalTL tl ascii st rt dt = evalTL' =<< liftSafe (parseTL ascii tl) where evalTL' il = evalIL il st rt dt
+runTL :: BIO m => TokenList -> Bool -> StackType -> RAMType -> DumpType -> m ()
+runTL tl ascii st rt dt = runTL' =<< liftSafe (parseTL ascii tl) where runTL' il = runIL il st rt dt
 
-evalIL :: BIO m => InstructionList -> StackType -> RAMType -> DumpType -> m ()
-evalIL il s ListRAMType    = evalIL' il s []
-evalIL il s SeqRAMType     = evalIL' il s Seq.empty
-evalIL il s SListRAMType   = evalIL' il s SList.sListEmpty
-evalIL il s MapListRAMType = evalIL' il s MapList.mapListEmpty
+runIL :: BIO m => InstructionList -> StackType -> RAMType -> DumpType -> m ()
+runIL il s ListRAMType    = runIL' il s []
+runIL il s SeqRAMType     = runIL' il s Seq.empty
+runIL il s SListRAMType   = runIL' il s SList.sListEmpty
+runIL il s MapListRAMType = runIL' il s MapList.mapListEmpty
 
-evalIL' :: (REvaluator Symbol r m) => InstructionList -> StackType -> r -> DumpType -> m ()
-evalIL' il ListStackType  = start il []
-evalIL' il SeqStackType   = start il Seq.empty
-evalIL' il SListStackType = start il SList.sListEmpty
+runIL' :: (REvaluator Symbol r m) => InstructionList -> StackType -> r -> DumpType -> m ()
+runIL' il ListStackType  = start il []
+runIL' il SeqStackType   = start il Seq.empty
+runIL' il SListStackType = start il SList.sListEmpty
 
 start :: (SREvaluator Symbol s r m) => InstructionList -> s -> r -> DumpType -> m ()
 start il s r dt = logDump dt =<< next (CU (Vector.fromList il) 0 (IS [])) s r